backend/requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
python-multipart
python-dotenv
